baiduface/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.7
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
